Move Kiezkassen Process to subpackage
avoids cyclic imports
