+++ src/meinberlin/meinberlin/static/js/Packages/MeinBerlin/Kiezkassen/Kiezkassen.ts
@@ -1,62 +1,16 @@
-import AdhConfig = require("../../Config/Config");
-import AdhHttp = require("../../Http/Http");
-import AdhMovingColumns = require("../../MovingColumns/MovingColumns");
-import AdhTabs = require("../../Tabs/Tabs");
-
+import AdhMeinBerlinKiezkassenProcess = require("./Process/Process");
 import AdhMeinBerlinKiezkassenProposal = require("./Proposal/Proposal");
-
-var pkgLocation = "/MeinBerlin/Kiezkassen";
-
-
-export var detailDirective = (adhConfig : AdhConfig.IService, adhHttp : AdhHttp.Service<any>) => {
-    return {
-        restrict: "E",
-        templateUrl: adhConfig.pkg_path + pkgLocation + "/Detail.html",
-        scope: {
-            path: "@"
-        },
-        require: "^adhMovingColumn",
-        link: (scope, element, attrs, column : AdhMovingColumns.MovingColumnController) => {
-            scope.currentTab = 0;
-            scope.tabs = [{
-                heading: "Tab1",
-                content: "foo1"
-            }, {
-                heading: "Tab2",
-                content: "foo2"
-            }];
-            scope.$watch(() => column.$scope.shared.isShowMap, function(value) {
-                scope.showMap = (typeof value === "undefined") ? true : value;
-            });
-            scope.$watch("path", (value : string) => {
-                if (value) {
-                    adhHttp.get(value).then((resource) => {
-                        // FIXME: set individual fields on scope, not simply dump whole resource
-                        scope.resource = resource;
-
-                        // FIXME: get actual polygon for this process
-                        scope.polygon = [[ 13.416490486705714, 52.52696152386536 ], [ 13.416354867070547, 52.527021169051629 ], [ 13.416329251729895, 52.527028815980323 ], [ 13.416223352178447, 52.527060448947488 ], [ 13.416145694406717, 52.52708363980976 ], [ 13.415954501339767, 52.527185233566222 ], [ 13.415932786518473, 52.527196778516675 ], [ 13.41591829520166, 52.52720448111279 ], [ 13.415902843699524, 52.527212694228353 ], [ 13.415888263794459, 52.527220432609056 ], [ 13.41569138227179, 52.527325053248823 ], [ 13.415431659144444, 52.527463063130241 ], [ 13.415212525654788, 52.52752488539879 ], [ 13.415144182754483, 52.527544174126973 ], [ 13.414861898951376, 52.527623810944405 ], [ 13.414451691419401, 52.52773955022306 ], [ 13.414394222095112, 52.527755757670121 ], [ 13.414317497860406, 52.527777403325452 ], [ 13.414236037281254, 52.527800388405723 ], [ 13.414183894186873, 52.527815104564077 ], [ 13.413582992224674, 52.527984633025746 ], [ 13.413509557519108, 52.528005349482932 ], [ 13.41338639893358, 52.528040094098991 ], [ 13.413376041042913, 52.528043014020149 ], [ 13.413330669853528, 52.528055818894117 ], [ 13.413316549461291, 52.528059801422074 ], [ 13.413244310880026, 52.52807992676113 ], [ 13.413136690528189, 52.528109927301074 ], [ 13.413104982817677, 52.528118757744046 ], [ 13.413096130040032, 52.52812122204476 ], [ 13.413085137783055, 52.528124293566258 ], [ 13.413051334863535, 52.528133713323108 ], [ 13.412888295676989, 52.528179151266009 ], [ 13.41280206922352, 52.528203178170962 ], [ 13.412269850337195, 52.528351499073857 ], [ 13.412230823865517, 52.528362373934002 ], [ 13.41218472979655, 52.528375221991048 ], [ 13.411936711720273, 52.528445127616038 ], [ 13.411860898450202, 52.528466809277191 ], [ 13.411771746207004, 52.528492303908536 ], [ 13.411692627592352, 52.528514923043296 ], [ 13.411678536154513, 52.528518950360912 ], [ 13.411626877291638, 52.528533729177518 ], [ 13.411575513527508, 52.528548418642821 ], [ 13.411542933467693, 52.528557732353264 ], [ 13.411483852536051, 52.528574627418244 ], [ 13.411433595319089, 52.528589004306099 ], [ 13.411425435544697, 52.528591334944565 ], [ 13.411313731971267, 52.528621164641415 ], [ 13.411278002259026, 52.528626707196644 ], [ 13.411233071021078, 52.528633670831681 ], [ 13.411069419043777, 52.528659056235568 ], [ 13.410863386618441, 52.52869100516569 ], [ 13.410824206169972, 52.528697089468146 ], [ 13.410809312644185, 52.528699389808182 ], [ 13.410801674130443, 52.52870058006701 ], [ 13.410273098751713, 52.528782551189302 ], [ 13.410091779757279, 52.528810670779201 ], [ 13.409994639077141, 52.528830254296352 ], [ 13.409968473291878, 52.528835535386058 ], [ 13.409889961252114, 52.528851360620543 ], [ 13.409811463846433, 52.528867194816257 ], [ 13.409785298061511, 52.528872466878965 ], [ 13.409714337651639, 52.528886778129895 ], [ 13.40939701387078, 52.528950764028295 ], [ 13.409305229269014, 52.528971731750659 ], [ 13.409270223644429, 52.528979728348638 ], [ 13.409056147653789, 52.529028634882948 ], [ 13.409013426799547, 52.529038396546554 ], [ 13.409002481033031, 52.529040901623794 ], [ 13.408743750978068, 52.529100015006193 ], [ 13.408573840510355, 52.529138838247235 ], [ 13.408394459137462, 52.529179818511565 ], [ 13.408189183983266, 52.529227526531415 ], [ 13.408157624445318, 52.529232860167674 ], [ 13.408005519108452, 52.529258577184521 ], [ 13.407691002884311, 52.529311051970446 ], [ 13.407505091999523, 52.529343221005178 ], [ 13.407438815269856, 52.529354690840911 ], [ 13.407414687800756, 52.529358869800873 ], [ 13.407232231803402, 52.529389580006352 ], [ 13.407207190546695, 52.529393793175593 ], [ 13.407112321119593, 52.52940949636465 ], [ 13.406870597712265, 52.529449504914147 ], [ 13.406869984855362, 52.529448299582761 ], [ 13.406765484154631, 52.529465799879866 ], [ 13.406732095828612, 52.529471390346458 ], [ 13.406625191232267, 52.529489281458446 ], [ 13.406625540964173, 52.52949006393559 ], [ 13.406512472855468, 52.529508779230305 ], [ 13.406478258943151, 52.529514439981483 ], [ 13.406366724457801, 52.529532897381728 ], [ 13.406184692838679, 52.529563903317523 ], [ 13.406145154039406, 52.52957064141907 ], [ 13.406133901513305, 52.529572552549347 ], [ 13.406042332511593, 52.529588153193494 ], [ 13.405900414485306, 52.529612322844713 ], [ 13.405618425627495, 52.529659514813453 ], [ 13.405585744879675, 52.529664989453799 ], [ 13.405594829594875, 52.529683967973149 ], [ 13.405802832781747, 52.530118405045158 ], [ 13.405895709261175, 52.530312399496438 ], [ 13.405916227007348, 52.530355267499687 ], [ 13.405928374468449, 52.530380623142868 ], [ 13.40595068711117, 52.530412315264556 ], [ 13.406057627351476, 52.530564198046839 ], [ 13.406063362767329, 52.53057235950709 ], [ 13.406086421839703, 52.530605104415592 ], [ 13.406044673919077, 52.530681960129925 ], [ 13.40601897358404, 52.530729289109978 ], [ 13.405991057243854, 52.530780675855645 ], [ 13.406045951425932, 52.530953292305746 ], [ 13.406051797627022, 52.530971698597348 ], [ 13.406061736109875, 52.531008588788318 ], [ 13.406089642233946, 52.531112138477745 ], [ 13.406097815781713, 52.531142492180145 ], [ 13.406102652885307, 52.531159377867048 ], [ 13.40613409324806, 52.531269314554507 ], [ 13.406135483551164, 52.531274169862286 ], [ 13.40614050886948, 52.531291729886419 ], [ 13.406160030887337, 52.531359946935645 ], [ 13.406196294289751, 52.531486706354208 ], [ 13.406203173323529, 52.531510758090455 ], [ 13.406205085123773, 52.531517411671231 ], [ 13.406206750469067, 52.531523256006686 ], [ 13.406212195752071, 52.531542299583911 ], [ 13.40625108089729, 52.53167819416182 ], [ 13.406271327219164, 52.531748973692928 ], [ 13.406277612562681, 52.531770948433781 ], [ 13.406297424544798, 52.531840199443351 ], [ 13.406347896182231, 52.532016608778861 ], [ 13.406362291989446, 52.532066906124726 ], [ 13.40636754921352, 52.532085275354461 ], [ 13.406379545729378, 52.532140933178113 ], [ 13.406411804835612, 52.532290717486219 ], [ 13.406424247844276, 52.532348461004858 ], [ 13.406439596955506, 52.532364368735784 ], [ 13.406475807305238, 52.532401918855534 ], [ 13.406521913368247, 52.532449714022924 ], [ 13.406538596650565, 52.532466999158807 ], [ 13.406574147552906, 52.532503847073201 ], [ 13.406689039525173, 52.532622961224931 ], [ 13.406799416831484, 52.532737375915858 ], [ 13.406912917356149, 52.532855031444704 ], [ 13.406936242054616, 52.532879212522879 ], [ 13.40700178894223, 52.532947146417989 ], [ 13.407003108335731, 52.532948523803135 ], [ 13.407005116876077, 52.532950594428776 ], [ 13.407026433133131, 52.532972704863248 ], [ 13.407106948633267, 52.533056168209832 ], [ 13.407161867135793, 52.53311308280589 ], [ 13.407179899617784, 52.53313178123372 ], [ 13.4072049925482, 52.533159002977413 ], [ 13.407299443013123, 52.53326148052642 ], [ 13.407335997977977, 52.533301142844735 ], [ 13.407350570845912, 52.533316222256829 ], [ 13.407476844876777, 52.53344691336887 ], [ 13.407482723941103, 52.533452990151837 ], [ 13.407499657289517, 52.53347051824764 ], [ 13.407579970030206, 52.53355363041463 ], [ 13.407660590973126, 52.533637057623913 ], [ 13.407677304579373, 52.533654351637736 ], [ 13.407683300919601, 52.533660563425137 ], [ 13.408180947047885, 52.5341755474903 ], [ 13.408222705691353, 52.534218166709266 ], [ 13.408309405108728, 52.534306628143426 ], [ 13.408349873729957, 52.534347923919448 ], [ 13.408301294665852, 52.534367074400343 ], [ 13.408165258228964, 52.534420715383185 ], [ 13.408115940698439, 52.534440160979116 ], [ 13.407803567442405, 52.534563983885683 ], [ 13.407675191934841, 52.534614870628566 ], [ 13.407649927275346, 52.534624888760206 ], [ 13.407608567696821, 52.534641275487026 ], [ 13.407565746195502, 52.534658252602128 ], [ 13.407535121436549, 52.534670390587188 ], [ 13.407486171945799, 52.534689800655791 ], [ 13.407446246563465, 52.534761421054533 ], [ 13.407361354161706, 52.534913773770604 ], [ 13.407345696121267, 52.534941863531749 ], [ 13.40732316808413, 52.534982288017368 ], [ 13.407298008058426, 52.535027425535475 ], [ 13.407286959545257, 52.53504726523353 ], [ 13.407262141440681, 52.535091792295013 ], [ 13.407190363275459, 52.535220606574825 ], [ 13.4071808464663, 52.535237663286182 ], [ 13.407177188365166, 52.535244234624003 ], [ 13.40712951447213, 52.535329778609253 ], [ 13.407119670353092, 52.535347445790663 ], [ 13.407114807776608, 52.535356171653298 ], [ 13.407051802308519, 52.535469230801461 ], [ 13.406976409452572, 52.535604517506698 ], [ 13.406917552174349, 52.5357101254803 ], [ 13.406882127944932, 52.535791917773835 ], [ 13.406855142141469, 52.535854242986318 ], [ 13.406838920950712, 52.535891677596013 ], [ 13.406782339967524, 52.536022335116243 ], [ 13.406778866005316, 52.536030371583479 ], [ 13.406748166815699, 52.536101397792805 ], [ 13.406712224879357, 52.536171650396099 ], [ 13.406707804652832, 52.53618029618459 ], [ 13.4066919246852, 52.536211333014997 ], [ 13.406591598740162, 52.536407483639643 ], [ 13.4065773705009, 52.53643531534312 ], [ 13.40656198149153, 52.536465391512024 ], [ 13.406546175678814, 52.536496293661877 ], [ 13.406531798569489, 52.536524412650749 ], [ 13.40649959149531, 52.536587384077379 ], [ 13.406483309455766, 52.536619192969212 ], [ 13.406466431914463, 52.536652195955689 ], [ 13.406419624134092, 52.536743726260184 ], [ 13.406404428339888, 52.536773434320111 ], [ 13.406401526148809, 52.536779099402445 ], [ 13.406337959523068, 52.536903381245267 ], [ 13.406324445473178, 52.536929794378366 ], [ 13.406255699098521, 52.537064176346448 ], [ 13.40623682675187, 52.537101093698148 ], [ 13.406175982880871, 52.537220051522965 ], [ 13.406156143132016, 52.537258827264701 ], [ 13.406143328332952, 52.537283884716594 ], [ 13.406088011690484, 52.537393812370219 ], [ 13.406068734205231, 52.537432112859648 ], [ 13.406031980241147, 52.537505158530976 ], [ 13.406020398669222, 52.537528196304677 ], [ 13.405949644281533, 52.537668810971851 ], [ 13.405939104725975, 52.537689765811798 ], [ 13.40590479172139, 52.537757954295621 ], [ 13.405885171384464, 52.537796964047146 ], [ 13.405814035785774, 52.537784331088183 ], [ 13.405782096174466, 52.537849414634891 ], [ 13.405762902633874, 52.53788851502734 ], [ 13.405745793564886, 52.537923395651539 ], [ 13.405504954285792, 52.538414210920131 ], [ 13.405483318142673, 52.538458294207501 ], [ 13.405461577660294, 52.538502601958079 ], [ 13.405518779714336, 52.538513699357338 ], [ 13.405415294074375, 52.53872061782711 ], [ 13.405363048170571, 52.53871158755544 ], [ 13.405302504362437, 52.538825647820332 ], [ 13.405357201219145, 52.538836758583692 ], [ 13.405322779989984, 52.53890559371262 ], [ 13.405249560166968, 52.539051999613754 ], [ 13.405177977577967, 52.539195128457813 ], [ 13.405134154598082, 52.53928699630552 ], [ 13.405122863637622, 52.539310672570721 ], [ 13.405100743410632, 52.539357064406467 ], [ 13.405082332223929, 52.539395671886751 ], [ 13.405068970851428, 52.539423657789939 ], [ 13.405022659667019, 52.5395207690501 ], [ 13.404990707927141, 52.539587775454294 ], [ 13.404980742703312, 52.539608641413651 ], [ 13.404946377667072, 52.539680711670293 ], [ 13.404914485282196, 52.539747592352491 ], [ 13.404905666941618, 52.539766061058572 ], [ 13.404879598926501, 52.539820704045347 ], [ 13.404870065362244, 52.539840708099604 ], [ 13.404838977187918, 52.539905891818655 ], [ 13.404828892461577, 52.539927045108179 ], [ 13.404795897465045, 52.539996224232901 ], [ 13.404704553776739, 52.540187717005537 ], [ 13.404607383699725, 52.540228108761426 ], [ 13.404475622633239, 52.540282876802259 ], [ 13.404190771882, 52.540401268983523 ], [ 13.403875667777378, 52.540311619995876 ], [ 13.403668000171168, 52.540253086217405 ], [ 13.403524797518202, 52.540212556775131 ], [ 13.403457537796742, 52.54019355828752 ], [ 13.403201893251795, 52.540121201819197 ], [ 13.403111380283226, 52.540095581106328 ], [ 13.403021588878815, 52.540070105471642 ], [ 13.402858534091886, 52.540023840262414 ], [ 13.402506052671695, 52.539923692586328 ], [ 13.402153424094324, 52.539823813173385 ], [ 13.402077982571427, 52.539802524792947 ], [ 13.401984350395409, 52.539776106484261 ], [ 13.401800059275647, 52.539724111038034 ], [ 13.40173636269391, 52.539706106854915 ], [ 13.401588321320613, 52.53966426282382 ], [ 13.401522549737521, 52.539645679442813 ], [ 13.401354329242364, 52.539598557596939 ], [ 13.401168139140422, 52.539546404778086 ], [ 13.400965543716666, 52.539488855438108 ], [ 13.40073298741121, 52.539422315973098 ], [ 13.400613692572493, 52.539388182095649 ], [ 13.400563316262001, 52.539373878201161 ], [ 13.400161887079808, 52.539261704909691 ], [ 13.399911401750845, 52.539191292273095 ], [ 13.399763347945745, 52.539149670591691 ], [ 13.399511685667964, 52.539079093093193 ], [ 13.399449329395383, 52.539061620452109 ], [ 13.399387326332517, 52.539044256294417 ], [ 13.399259037758821, 52.539008315461096 ], [ 13.399195313310496, 52.538990399718102 ], [ 13.399045274131286, 52.538948207170236 ], [ 13.398989129045869, 52.538932426708278 ], [ 13.398978488733082, 52.538929431705085 ], [ 13.398831835140884, 52.538888135063246 ], [ 13.398769891937359, 52.538870689817962 ], [ 13.398713732497542, 52.538854873251346 ], [ 13.398524915946144, 52.53880166876958 ], [ 13.397963177275617, 52.538643590321684 ], [ 13.397786870727636, 52.538594093300773 ], [ 13.397610124773665, 52.538544145840888 ], [ 13.397441767727825, 52.538496550856273 ], [ 13.397273258980597, 52.538449853987778 ], [ 13.397105196950653, 52.538402348959956 ], [ 13.396903580537547, 52.538345612318409 ], [ 13.396805865857411, 52.538317564404821 ], [ 13.396623680358038, 52.538265448141388 ], [ 13.396505255454928, 52.538232444101801 ], [ 13.396336264382848, 52.538185358537845 ], [ 13.395858913273306, 52.538084103180616 ], [ 13.39568956100595, 52.538035874700689 ], [ 13.395630108257487, 52.538018944899193 ], [ 13.395543828260861, 52.537994378067907 ], [ 13.395188084406838, 52.537893069866769 ], [ 13.394996748007797, 52.537838587715179 ], [ 13.394797862911531, 52.537781942688895 ], [ 13.394607101398384, 52.53772762278907 ], [ 13.394527326917407, 52.537704910204695 ], [ 13.394395561666617, 52.537667384627191 ], [ 13.394213202293979, 52.537615749573163 ], [ 13.394015377001953, 52.537559734365296 ], [ 13.393891278416346, 52.537524595733885 ], [ 13.393768622182524, 52.537489864203238 ], [ 13.393517497533445, 52.537418762751543 ], [ 13.393264156466081, 52.537282216661595 ], [ 13.393227592041905, 52.53726276963522 ], [ 13.393075640210895, 52.537166619731856 ], [ 13.392992737875392, 52.537114370277557 ], [ 13.392901007972421, 52.53705654966685 ], [ 13.39275569475106, 52.536964959633615 ], [ 13.392698747613753, 52.53692907177912 ], [ 13.392579800894868, 52.536854095783397 ], [ 13.392457597884702, 52.536776291459368 ], [ 13.392347454157195, 52.536703363599635 ], [ 13.392220039961396, 52.536624605126065 ], [ 13.392155752398946, 52.536583562183317 ], [ 13.392089555612548, 52.536541293255837 ], [ 13.391964285829079, 52.536461307603247 ], [ 13.391931841186302, 52.536440592259382 ], [ 13.391904037846476, 52.536422833688214 ], [ 13.391766798461338, 52.536335212507218 ], [ 13.39148647593187, 52.536156210675735 ], [ 13.39148524222967, 52.536155417411308 ], [ 13.391321714431312, 52.536051001598608 ], [ 13.391320568821854, 52.536050271413224 ], [ 13.391314591107712, 52.536046458232349 ], [ 13.391280399309482, 52.536024624906304 ], [ 13.391278357808579, 52.536023317802005 ], [ 13.391277564691634, 52.536022812980008 ], [ 13.391072458697902, 52.535891839857726 ], [ 13.390864871778422, 52.535759288793741 ], [ 13.390649208337832, 52.535621571984855 ], [ 13.390599108064377, 52.53559018103202 ], [ 13.390534995736473, 52.535550000253629 ], [ 13.390424528902601, 52.535480772380971 ], [ 13.390374164319702, 52.535457495621642 ], [ 13.390332455271592, 52.535438819246295 ], [ 13.390198169023586, 52.535378709828088 ], [ 13.390149388542552, 52.535356873979303 ], [ 13.390059152630601, 52.535277468533941 ], [ 13.390006701027811, 52.535231298805563 ], [ 13.389812193625559, 52.535060148013969 ], [ 13.389758407774869, 52.535012807256706 ], [ 13.389631724508275, 52.534901345520886 ], [ 13.389502988612072, 52.534788046285172 ], [ 13.389455466332661, 52.534746226670762 ], [ 13.388340038800488, 52.53376460862377 ], [ 13.388236335939553, 52.533620705404012 ], [ 13.388187465075253, 52.533553540590574 ], [ 13.388119841917106, 52.533460591599784 ], [ 13.387956526454436, 52.533236120687917 ], [ 13.387730833392622, 52.53313926689173 ], [ 13.387671647853605, 52.533183450884579 ], [ 13.387623707152759, 52.533219300105699 ], [ 13.387601586324381, 52.533235844522338 ], [ 13.38755848404727, 52.533268063953543 ], [ 13.387259414705715, 52.533491663796632 ], [ 13.387240726470237, 52.53350563596053 ], [ 13.387046504406451, 52.533650815360211 ], [ 13.386860271359222, 52.53379002561114 ], [ 13.386835116689841, 52.533808819350028 ], [ 13.386377681758129, 52.534150744808585 ], [ 13.386363698479139, 52.534161194687186 ], [ 13.386320845940373, 52.534193225454565 ], [ 13.38630687742004, 52.534203666369848 ], [ 13.386192554288929, 52.534289120499956 ], [ 13.384897756164525, 52.53542013063656 ], [ 13.384636489117963, 52.535630330239009 ], [ 13.384418873220961, 52.535805409478442 ], [ 13.384343718995149, 52.535865878604575 ], [ 13.384319679116127, 52.535885213277339 ], [ 13.38393823515389, 52.536192084660172 ], [ 13.383511681821947, 52.53653226497805 ], [ 13.383475711037599, 52.536560956877949 ], [ 13.383449613754191, 52.536581760931867 ], [ 13.383438644870992, 52.536590509220325 ], [ 13.383358208422775, 52.536659602892726 ], [ 13.383037393041695, 52.536935170159403 ], [ 13.382611923277633, 52.537300628479535 ], [ 13.382517386574891, 52.537381842253161 ], [ 13.38236934553791, 52.537509190079298 ], [ 13.382217026114697, 52.537639826929066 ], [ 13.382101163576026, 52.537739346761576 ], [ 13.382095388223043, 52.53774430436458 ], [ 13.381124111669667, 52.538679558772905 ], [ 13.380637919565824, 52.539147249193419 ], [ 13.38001019771623, 52.539754212095318 ], [ 13.379951930320514, 52.539828813421863 ], [ 13.379870801115993, 52.539932663048575 ], [ 13.379835915068668, 52.539977253217316 ], [ 13.379781282985318, 52.540047413720167 ], [ 13.379710749216922, 52.540114665335082 ], [ 13.37966725984054, 52.540111213625593 ], [ 13.379547443087951, 52.540101707400865 ], [ 13.379515400650902, 52.540099160214432 ], [ 13.379513456368739, 52.540104188596089 ], [ 13.379460133121746, 52.540072444733028 ], [ 13.379014605527457, 52.539806330525934 ], [ 13.377820600539705, 52.539094682875024 ], [ 13.377609830435484, 52.53896904849519 ], [ 13.377398854537695, 52.538843530127394 ], [ 13.376542654285833, 52.53833408489534 ], [ 13.376127501729236, 52.538087102262168 ], [ 13.37616706871467, 52.538062041772243 ], [ 13.375900706151377, 52.53790460020975 ], [ 13.375717103527361, 52.53779547428416 ], [ 13.375581119858802, 52.5377147925564 ], [ 13.37548605791104, 52.537658389211238 ], [ 13.37520544190005, 52.537491821470354 ], [ 13.374925856148606, 52.537325866331969 ], [ 13.374712799395127, 52.537159791575817 ], [ 13.374650251108102, 52.537111122260953 ], [ 13.374546148533124, 52.537033122292641 ], [ 13.374318544661703, 52.537173536662173 ], [ 13.374175332945942, 52.537261721355542 ], [ 13.373947268756096, 52.537402322742089 ], [ 13.373750545051367, 52.537523373121175 ], [ 13.373743283068594, 52.537527841869021 ], [ 13.373699681665526, 52.537554654283618 ], [ 13.373428786447112, 52.537736150569657 ], [ 13.373205292437115, 52.537885899576757 ], [ 13.373002707530105, 52.538023327543208 ], [ 13.372982200394404, 52.538037239632665 ], [ 13.372344793789141, 52.53773904721843 ], [ 13.372158509114971, 52.53765189998478 ], [ 13.371950677756423, 52.537554677139241 ], [ 13.369365121158381, 52.536345017452206 ], [ 13.368606673729381, 52.53599027700794 ], [ 13.367883263940941, 52.536621230904842 ], [ 13.367820699372896, 52.536591771095615 ], [ 13.367738294172183, 52.536552966806809 ], [ 13.3676610476983, 52.536516591121945 ], [ 13.367563459957653, 52.536470815895406 ], [ 13.367483216938485, 52.536432806974581 ], [ 13.36740253193847, 52.536394797030709 ], [ 13.36732169970472, 52.536356786708836 ], [ 13.367241014983325, 52.536318776654717 ], [ 13.367160330401171, 52.536280766545531 ], [ 13.367079646488387, 52.536242666517346 ], [ 13.366998962184887, 52.536204656298032 ], [ 13.366918130648289, 52.536166645700199 ], [ 13.3668374466234, 52.536128635370645 ], [ 13.366701944785445, 52.536064803158546 ], [ 13.366648330675929, 52.5360073514288 ], [ 13.366625176169531, 52.535982767375913 ], [ 13.366560326323096, 52.535913878046095 ], [ 13.366463258049057, 52.535877871385878 ], [ 13.366385554609971, 52.535849366042264 ], [ 13.365860330943041, 52.535660445540479 ], [ 13.365864185481209, 52.535656589834602 ], [ 13.365984649687855, 52.535539347607603 ], [ 13.366217245981957, 52.535312959264985 ], [ 13.366547192963427, 52.534991823988307 ], [ 13.366644916171788, 52.53489671879457 ], [ 13.366664510899684, 52.53487769246415 ], [ 13.36686647802898, 52.534681061769589 ], [ 13.367172937417475, 52.534382815728783 ], [ 13.3674918283133, 52.534072419381872 ], [ 13.367572502576992, 52.533993901124504 ], [ 13.367688140710548, 52.533881337533622 ], [ 13.367766650380972, 52.533804917249562 ], [ 13.367843099577369, 52.533730505385662 ], [ 13.367902029595871, 52.533673147517831 ], [ 13.368226586807317, 52.533357207886347 ], [ 13.368243705536621, 52.533340503391095 ], [ 13.368288331861242, 52.533297106097613 ], [ 13.368504264385022, 52.533264596840432 ], [ 13.368851225350818, 52.532926507661657 ], [ 13.368931568939832, 52.532848320248277 ], [ 13.368984488923651, 52.532795145449548 ], [ 13.369055873512817, 52.532721800147613 ], [ 13.369131507398983, 52.53264247902932 ], [ 13.369180402257669, 52.532587066725803 ], [ 13.369242244098919, 52.532522606183853 ], [ 13.369286838410648, 52.532474292815174 ], [ 13.369529254455863, 52.532203435975013 ], [ 13.369614776982989, 52.53210602819636 ], [ 13.369837560120949, 52.531847888918776 ], [ 13.369933978718233, 52.531731498066435 ], [ 13.36994046536706, 52.531723433259266 ], [ 13.369963843891314, 52.531694385629606 ], [ 13.369969484407376, 52.531687379394455 ], [ 13.369973373473442, 52.53168253511074 ], [ 13.370061334563758, 52.531573449818453 ], [ 13.370139582460046, 52.531469735314914 ], [ 13.3701814649221, 52.531413561506639 ], [ 13.370273052280302, 52.531288901289351 ], [ 13.370339526117549, 52.531197697332523 ], [ 13.370409026207328, 52.531100451963319 ], [ 13.370479177871912, 52.531000071667464 ], [ 13.370484928740517, 52.5309917985468 ], [ 13.370505643627494, 52.530962017132779 ], [ 13.370511171623193, 52.530954058056516 ], [ 13.370514619179827, 52.530949095974059 ], [ 13.370589551778327, 52.530842435395805 ], [ 13.370686659971591, 52.53070556512867 ], [ 13.37076056031529, 52.530606495814382 ], [ 13.370832462069949, 52.530508967813169 ], [ 13.370879288299841, 52.530448679625458 ], [ 13.370899219099012, 52.530454842523895 ], [ 13.370936874131415, 52.530408194000543 ], [ 13.370973009912911, 52.530366862201994 ], [ 13.371014955420081, 52.530319764611107 ], [ 13.37106819528155, 52.530261296493109 ], [ 13.371111717361453, 52.530214184299446 ], [ 13.371115749714669, 52.530209996302524 ], [ 13.371172211081898, 52.530149899546977 ], [ 13.37119888238305, 52.530121937158455 ], [ 13.371207436707412, 52.530112978091687 ], [ 13.37123784454641, 52.530081006785835 ], [ 13.371279643458031, 52.530038761503768 ], [ 13.371357009598592, 52.529959074224998 ], [ 13.371670286903695, 52.5296455629665 ], [ 13.371893823887554, 52.529423502236028 ], [ 13.372109340773163, 52.529209421828696 ], [ 13.372154165582938, 52.529165987534526 ], [ 13.372195259429496, 52.529128152789468 ], [ 13.372223414390515, 52.529103311679556 ], [ 13.37225123476402, 52.529080267147847 ], [ 13.372232505661268, 52.529072768067429 ], [ 13.372344857980522, 52.528976278662959 ], [ 13.372374942969591, 52.528951450650943 ], [ 13.372873336189599, 52.528553420166212 ], [ 13.373028260101046, 52.528427869122773 ], [ 13.373115922111419, 52.528355463854631 ], [ 13.373201849083117, 52.528282371828951 ], [ 13.373226034710701, 52.528260882935321 ], [ 13.373243643752563, 52.528245355990286 ], [ 13.373285306050839, 52.528208285934092 ], [ 13.373325748103174, 52.52817072797847 ], [ 13.373365896636539, 52.528132953704031 ], [ 13.373404001136743, 52.528094456119334 ], [ 13.373547545429638, 52.527946072239025 ], [ 13.373564573913287, 52.527952030780241 ], [ 13.373572797740488, 52.527941453258912 ], [ 13.373582075875689, 52.527929458121207 ], [ 13.373583869736777, 52.527927563257066 ], [ 13.373585913778117, 52.527925765103028 ], [ 13.373588077807606, 52.52792401995432 ], [ 13.373590470505551, 52.527922389629175 ], [ 13.37359297153265, 52.527920820945923 ], [ 13.37359567621133, 52.527919383439631 ], [ 13.373598475899684, 52.527918015266955 ], [ 13.373601451101141, 52.527916792613929 ], [ 13.373602950295497, 52.527916467659992 ], [ 13.373604490214127, 52.527916223614739 ], [ 13.373606046120649, 52.527916020766945 ], [ 13.373607625090273, 52.527915901377263 ], [ 13.373609211095596, 52.527915824105278 ], [ 13.373610802108365, 52.527915831413338 ], [ 13.373612391084297, 52.527915881038844 ], [ 13.373613966926362, 52.527916014919604 ], [ 13.373615531693897, 52.527916190593807 ], [ 13.373617065410908, 52.527916448757274 ], [ 13.373618579205122, 52.527916747475309 ], [ 13.37362004456325, 52.527917125505709 ], [ 13.373621481491094, 52.527917542158043 ], [ 13.373622853425784, 52.527918033589238 ], [ 13.373645167092922, 52.527924785247862 ], [ 13.373654310489725, 52.5279157194229 ], [ 13.373662850456467, 52.527919350264078 ], [ 13.373665239391121, 52.52791645272471 ], [ 13.373668088244971, 52.527913007990691 ], [ 13.373695504226351, 52.527875458003408 ], [ 13.373748731975322, 52.527802574251737 ], [ 13.373824251668344, 52.527699138996091 ], [ 13.37383716092959, 52.527683440162335 ], [ 13.373817450904109, 52.527677466943807 ], [ 13.373814867545436, 52.5276766979146 ], [ 13.373812398730788, 52.527675799769256 ], [ 13.373809993745773, 52.527674839361509 ], [ 13.373807728930878, 52.527673759609762 ], [ 13.37380554023529, 52.527672622964765 ], [ 13.373803515126047, 52.527671378629442 ], [ 13.373801577225711, 52.527670083666983 ], [ 13.373799823748936, 52.527668694365651 ], [ 13.373798167195771, 52.527667261501456 ], [ 13.373796712995826, 52.527665749137107 ], [ 13.373795230731423, 52.527664038112675 ], [ 13.373794014030729, 52.527662250875942 ], [ 13.373792934191016, 52.52766043122525 ], [ 13.373792134650397, 52.527658558505337 ], [ 13.373791478007865, 52.527656665232477 ], [ 13.373791111007767, 52.527654743054391 ], [ 13.373790890193956, 52.527652812579753 ], [ 13.373790962794144, 52.52765087791991 ], [ 13.373791182056676, 52.527648947379511 ], [ 13.373791692860445, 52.527647037454422 ], [ 13.373800391612344, 52.527620455933551 ], [ 13.373765347542493, 52.527616247243174 ], [ 13.373788170230396, 52.527542067332824 ], [ 13.373806214360478, 52.527478930630473 ], [ 13.37382144704719, 52.527417854818069 ], [ 13.373834795719752, 52.527351293198222 ], [ 13.373841644241645, 52.527287683241546 ], [ 13.373843505810324, 52.527220108558652 ], [ 13.373840909427527, 52.527159084513123 ], [ 13.373834463576939, 52.527101287390906 ], [ 13.373818806957084, 52.527030979318518 ], [ 13.373796652902966, 52.526963173585742 ], [ 13.373760488063908, 52.526897674401084 ], [ 13.373718253943562, 52.52683710481967 ], [ 13.373674824167598, 52.526779498219362 ], [ 13.373622838924973, 52.526723580747586 ], [ 13.373565827182855, 52.526670618054645 ], [ 13.373504969221885, 52.526620343062604 ], [ 13.373436293577269, 52.526571579018679 ], [ 13.373367155897199, 52.526526318690117 ], [ 13.373040115219933, 52.526526409005498 ], [ 13.371431226704274, 52.526526912774081 ], [ 13.371051054342262, 52.526527026765088 ], [ 13.37100389451753, 52.526498374692714 ], [ 13.371001434193397, 52.526224343242852 ], [ 13.370996373460555, 52.525662359871866 ], [ 13.370993883834373, 52.52538577614596 ], [ 13.370992575453466, 52.525235599457986 ], [ 13.371318072125527, 52.525037753950386 ], [ 13.371338195581945, 52.525025530762548 ], [ 13.371372572989102, 52.525004567472962 ], [ 13.371641563160829, 52.524834637115895 ], [ 13.371665874269203, 52.524819268625784 ], [ 13.371720825658031, 52.524784519302301 ], [ 13.371720778654627, 52.524408477918769 ], [ 13.371720753286167, 52.524300918100884 ], [ 13.371720721149462, 52.524143644787578 ], [ 13.371720639017566, 52.523781298758827 ], [ 13.371720621818632, 52.523649430383912 ], [ 13.371720576955221, 52.523606294943683 ], [ 13.371768887404897, 52.523606255235727 ], [ 13.371767414932632, 52.523573028862636 ], [ 13.371736534818494, 52.523547467552966 ], [ 13.371696455769946, 52.523388023060726 ], [ 13.371696320590679, 52.523368117604861 ], [ 13.37169616125022, 52.523344752282689 ], [ 13.3716958953681, 52.523304168538615 ], [ 13.371692506003606, 52.522787894027665 ], [ 13.371692468771517, 52.52278160337562 ], [ 13.371694250862694, 52.522781706067505 ], [ 13.371803544046758, 52.522786066336913 ], [ 13.371912569394093, 52.52278834105347 ], [ 13.372019913687723, 52.522788329473073 ], [ 13.372047580044176, 52.522788694550478 ], [ 13.372315365761233, 52.522788740260189 ], [ 13.372369154092469, 52.522786474436622 ], [ 13.37239311703377, 52.522785231871751 ], [ 13.372450366724747, 52.522783117233715 ], [ 13.372519117973662, 52.522779292871128 ], [ 13.372587858678743, 52.522774740536974 ], [ 13.37267570691381, 52.522767820788047 ], [ 13.37272256566377, 52.522763877421184 ], [ 13.372777932095305, 52.522758757081732 ], [ 13.372859673068316, 52.522750628883657 ], [ 13.372931864395319, 52.522742965424754 ], [ 13.372989024312158, 52.522735961660821 ], [ 13.37306355547874, 52.52272620028954 ], [ 13.373142676618061, 52.522715037820277 ], [ 13.373214698712943, 52.522703482658393 ], [ 13.373253499017009, 52.522696879780398 ], [ 13.373256317003143, 52.522696184865922 ], [ 13.373297862960561, 52.522688653254171 ], [ 13.373346565547727, 52.522679177885294 ], [ 13.373428197604724, 52.522661765974981 ], [ 13.373469003295142, 52.522652291628191 ], [ 13.373520035221153, 52.522639972441837 ], [ 13.373552776861901, 52.522631388453725 ], [ 13.373622544325901, 52.522614750830826 ], [ 13.37367833536449, 52.522597220578859 ], [ 13.373772501787839, 52.522570075856308 ], [ 13.373851539413391, 52.522545154355662 ], [ 13.373955058087835, 52.522510813284491 ], [ 13.374024052712224, 52.522487586646989 ], [ 13.374083678830477, 52.522466640523696 ], [ 13.374135029361588, 52.522447509961609 ], [ 13.374209633336012, 52.52241721380036 ], [ 13.374232820101176, 52.522407405042927 ], [ 13.374277364024254, 52.522388259898072 ], [ 13.374364608838263, 52.522350199278058 ], [ 13.374460809051486, 52.52230734090076 ], [ 13.374529348912409, 52.522275953228693 ], [ 13.374598190170181, 52.522243370950235 ], [ 13.374678735782025, 52.522204478052743 ], [ 13.374765189413754, 52.522160376502917 ], [ 13.374825169684648, 52.522128889558083 ], [ 13.374892451020136, 52.522093347219162 ], [ 13.374953783276592, 52.522059850102615 ], [ 13.374988883045271, 52.522040675612516 ], [ 13.375141106410972, 52.521957244810949 ], [ 13.375169703038226, 52.521941678015665 ], [ 13.375266370174966, 52.521891576830015 ], [ 13.375322050623669, 52.52186217435905 ], [ 13.375368608321828, 52.521830748491148 ], [ 13.375419791216277, 52.521791554035019 ], [ 13.375463877615678, 52.521749272191919 ], [ 13.375483087515589, 52.521728859625071 ], [ 13.37553997600814, 52.521665823025934 ], [ 13.375554639161917, 52.521649579535016 ], [ 13.37563607964424, 52.521559338838031 ], [ 13.375852259201773, 52.521319811916612 ], [ 13.375875054690846, 52.521293297527613 ], [ 13.375894269956614, 52.521265747997425 ], [ 13.375911646354018, 52.521237747768176 ], [ 13.375925263725625, 52.521208983854962 ], [ 13.375936966434832, 52.521179908074771 ], [ 13.375944786398447, 52.521150351257312 ], [ 13.375981533025621, 52.520991807821289 ], [ 13.375992314682133, 52.520947221911179 ], [ 13.376002936528268, 52.520902621731175 ], [ 13.376013478439125, 52.52085801449541 ], [ 13.37602386046456, 52.520813393309474 ], [ 13.376034162517866, 52.5207687652277 ], [ 13.376044304612426, 52.52072412351562 ], [ 13.376054366698863, 52.520679475067716 ], [ 13.376064268754956, 52.520634813309478 ], [ 13.376068926627015, 52.520613668874205 ], [ 13.376074689802488, 52.520587278643909 ], [ 13.376081824354285, 52.52055443616414 ], [ 13.376088872378574, 52.520521586733516 ], [ 13.37609587712684, 52.520488733858947 ], [ 13.376102795320701, 52.520455874160795 ], [ 13.37611212839904, 52.520414807171043 ], [ 13.376124269699991, 52.520374014299328 ], [ 13.376137807492144, 52.520333385311361 ], [ 13.376154119774815, 52.520293137356518 ], [ 13.376171808935561, 52.520253106398457 ], [ 13.376192227927223, 52.520213561887338 ], [ 13.376219100278233, 52.520166196842204 ], [ 13.376251358440182, 52.520120099710397 ], [ 13.376286251705242, 52.520074722391605 ], [ 13.376326282152121, 52.520030951634766 ], [ 13.376351879347316, 52.520004667515728 ], [ 13.37637911004456, 52.519979001503977 ], [ 13.376407142630992, 52.519953658772202 ], [ 13.376436747744087, 52.519928990424603 ], [ 13.376474595236765, 52.519898439830932 ], [ 13.376513561174523, 52.519868417007977 ], [ 13.376553078935107, 52.519838663796605 ], [ 13.376593685270903, 52.519809461103144 ], [ 13.376634828181077, 52.519780539239072 ], [ 13.376677028554223, 52.519752190011594 ], [ 13.376719749641012, 52.519724132510689 ], [ 13.376763495871709, 52.519696669111283 ], [ 13.376862195718932, 52.519637638287151 ], [ 13.376860322381468, 52.519469280584666 ], [ 13.376860116056763, 52.519451118361459 ], [ 13.376859853513674, 52.519427285515768 ], [ 13.376857340732998, 52.51920285946202 ], [ 13.376851979614869, 52.518722284147145 ], [ 13.37685188782023, 52.518715094724293 ], [ 13.376849132554582, 52.518468507328876 ], [ 13.376849055492091, 52.518461317936655 ], [ 13.376843693914459, 52.517980850386976 ], [ 13.376881523262499, 52.517791404170602 ], [ 13.376892462302001, 52.51776914064888 ], [ 13.37696264267967, 52.51762615119781 ], [ 13.377193180206358, 52.517156578624167 ], [ 13.377196239438133, 52.517148353238746 ], [ 13.377198011324582, 52.517139988342102 ], [ 13.377199132723812, 52.517131581683856 ], [ 13.377198947374799, 52.517123147994653 ], [ 13.377198108396483, 52.517114729090579 ], [ 13.377195969540924, 52.51710639617928 ], [ 13.37719318705604, 52.517098134302408 ], [ 13.377189137700586, 52.51709006971295 ], [ 13.377184683846373, 52.517082473652962 ], [ 13.377179123011677, 52.517075154198736 ], [ 13.377173025090215, 52.517067995702348 ], [ 13.377165891731774, 52.517061202653991 ], [ 13.377158261933149, 52.517054613422026 ], [ 13.377149687290656, 52.517048471756851 ], [ 13.377140665987817, 52.517042573016788 ], [ 13.377130807817061, 52.51703719572798 ], [ 13.377148717778736, 52.517008175552299 ], [ 13.377162468424673, 52.516978322944716 ], [ 13.377174098809256, 52.516948136765613 ], [ 13.377181428729523, 52.516917452075262 ], [ 13.377186585615226, 52.51688660327644 ], [ 13.37718737260346, 52.516855598198973 ], [ 13.377188747429511, 52.516846980761663 ], [ 13.377189028302999, 52.51683832430178 ], [ 13.377188761082367, 52.516829667681023 ], [ 13.377187399849211, 52.516821049440686 ], [ 13.377185494166277, 52.516812469690102 ], [ 13.377182509077532, 52.516804005208357 ], [ 13.377179082323554, 52.51679310712602 ], [ 13.377174377505041, 52.516782387961946 ], [ 13.37716904233236, 52.516771779230773 ], [ 13.377162468225237, 52.516761432538431 ], [ 13.377155286990018, 52.516751237145215 ], [ 13.377146920430015, 52.516741383848839 ], [ 13.377145393453217, 52.516738680767737 ], [ 13.377143454532289, 52.516736077828035 ], [ 13.377141315879875, 52.516733533775927 ], [ 13.377138792659363, 52.516731124420957 ], [ 13.37713608537787, 52.516728790603942 ], [ 13.377133028644771, 52.516726623328637 ], [ 13.377129807225916, 52.516724546721264 ], [ 13.3771262785056, 52.516722665148066 ], [ 13.377122607789483, 52.516720887547798 ], [ 13.377118678107315, 52.516719329548557 ], [ 13.377114631977337, 52.516717886734511 ], [ 13.377110380429102, 52.516716683668264 ], [ 13.37707595737511, 52.51670834344381 ], [ 13.376964029705421, 52.516661979379364 ], [ 13.376950162029953, 52.51665522818088 ], [ 13.376773588066747, 52.516540097277741 ], [ 13.376753848667478, 52.516521938722313 ], [ 13.376708260158553, 52.516474618156096 ], [ 13.376679318678503, 52.516428653869866 ], [ 13.376653840397907, 52.516381932830456 ], [ 13.376652237325906, 52.516378883006041 ], [ 13.376629375747784, 52.516306133586795 ], [ 13.376671011334922, 52.516110395862967 ], [ 13.376696636507425, 52.516069650418395 ], [ 13.37672916338928, 52.516030781141609 ], [ 13.376763799928192, 52.515994220154973 ], [ 13.376804372422692, 52.515960006801357 ], [ 13.376847758373962, 52.515927103298225 ], [ 13.376896422652173, 52.515897073123469 ], [ 13.376934299035824, 52.515876818834982 ], [ 13.376975173891337, 52.515858864640464 ], [ 13.377017409638535, 52.515842123730891 ], [ 13.37706207373963, 52.515827920912962 ], [ 13.377107800270586, 52.515815037656679 ], [ 13.377155335101881, 52.515804878293864 ], [ 13.377219220173497, 52.515786950048856 ], [ 13.377228664312515, 52.515783701257355 ], [ 13.377237585918632, 52.515779942760631 ], [ 13.377246218263631, 52.515775940751368 ], [ 13.377254219326385, 52.515771477114491 ], [ 13.377278278423196, 52.515755635561774 ], [ 13.37730009055028, 52.515738626291729 ], [ 13.377320721087694, 52.515721076941368 ], [ 13.377338883361121, 52.515702540064559 ], [ 13.377359308298692, 52.515670929016075 ], [ 13.37737415915317, 52.515638180894037 ], [ 13.377369728219241, 52.515619443606539 ], [ 13.377321637165377, 52.515415788123008 ], [ 13.377137446911819, 52.514635970065186 ], [ 13.377098880263262, 52.514472738038783 ], [ 13.376998287068723, 52.514046797315068 ], [ 13.376889108138307, 52.513584523723118 ], [ 13.376774896943813, 52.513100950238737 ], [ 13.376736879985048, 52.512939956860393 ], [ 13.376696518462788, 52.512769082298185 ], [ 13.3767764395933, 52.512546115190787 ], [ 13.376803351358104, 52.512470658080773 ], [ 13.376808171027426, 52.512457125534063 ], [ 13.376761367599745, 52.512243057020527 ], [ 13.376769191941738, 52.512239928255461 ], [ 13.376777952076585, 52.51223541753815 ], [ 13.376805761408967, 52.512219803732371 ], [ 13.376825125894742, 52.512194664321704 ], [ 13.376799713373135, 52.512078594340949 ], [ 13.376792491964325, 52.512045553531131 ], [ 13.376737950318811, 52.511796178479969 ], [ 13.376736012429003, 52.51178726871823 ], [ 13.376734447347838, 52.511780148066848 ], [ 13.376724124930126, 52.511732937902103 ], [ 13.376722761061822, 52.511726698357172 ], [ 13.376720004531229, 52.511714102381326 ], [ 13.376717506342267, 52.511702702161898 ], [ 13.376698871131003, 52.511617542216506 ], [ 13.376694779446517, 52.511598841553301 ], [ 13.376676130052529, 52.511513600693874 ], [ 13.376673631886113, 52.511502200473139 ], [ 13.376670904076244, 52.511489739354531 ], [ 13.376668793597224, 52.511480101316174 ], [ 13.376662462318693, 52.511451160241464 ], [ 13.376660681937805, 52.511443050613494 ], [ 13.376654652428247, 52.511415440184862 ], [ 13.376652872050215, 52.51140733055675 ], [ 13.376646870987997, 52.511379899918644 ], [ 13.376574126018996, 52.511047495521872 ], [ 13.37656966108336, 52.511027104592422 ], [ 13.376527510808176, 52.51083446071349 ], [ 13.376492339527463, 52.510673437262412 ], [ 13.376476116811515, 52.51059933582269 ], [ 13.37647343227767, 52.510587054519995 ], [ 13.376454539166446, 52.51050078862832 ], [ 13.376451868862663, 52.510488597220544 ], [ 13.376445494651044, 52.510459485293865 ], [ 13.376439335812359, 52.510431353356083 ], [ 13.37641093906587, 52.510301671308106 ], [ 13.376393110783633, 52.510219864983114 ], [ 13.376391704201376, 52.510213382705508 ], [ 13.376376830815529, 52.50900508257272 ], [ 13.376666064500638, 52.508769168055885 ], [ 13.376751826024357, 52.50869927211329 ], [ 13.376940187286088, 52.508545658319214 ], [ 13.37723311380074, 52.508306766663075 ], [ 13.377248287389053, 52.508295718233185 ], [ 13.37732302538792, 52.508234524575386 ], [ 13.377626414226119, 52.507986119252763 ], [ 13.377649156393712, 52.507967502051812 ], [ 13.37766183558003, 52.50795736494667 ], [ 13.377849241825523, 52.507807432168079 ], [ 13.377944643784314, 52.507730168680268 ], [ 13.378067831384085, 52.507630398267061 ], [ 13.378078321245466, 52.507621928015446 ], [ 13.378174372242166, 52.507544342188389 ], [ 13.378392246432494, 52.5073683396337 ], [ 13.37839717321407, 52.507364359942642 ], [ 13.378439117399495, 52.50733047878078 ], [ 13.378573234911631, 52.507222121649946 ], [ 13.378609630692832, 52.507192722068929 ], [ 13.378614010036189, 52.50718918156214 ], [ 13.378810220911502, 52.507030683548528 ], [ 13.378938134781595, 52.506930445835003 ], [ 13.378941447803586, 52.506930569597273 ], [ 13.37904789418973, 52.506934072525659 ], [ 13.379051855170177, 52.506934206627548 ], [ 13.379085972793586, 52.506935329455217 ], [ 13.379151498596082, 52.506937497520234 ], [ 13.379165310635271, 52.506937939796899 ], [ 13.379189577205851, 52.506938754416453 ], [ 13.379220131496375, 52.506939743939967 ], [ 13.379441859291441, 52.506947046244925 ], [ 13.37946404652544, 52.506948350723583 ], [ 13.382385268628827, 52.507119276716658 ], [ 13.382490314170882, 52.507125874015898 ], [ 13.382507789871239, 52.507126970495364 ], [ 13.384278124909141, 52.507238143813261 ], [ 13.385419661875126, 52.507309835246076 ], [ 13.385485857601314, 52.507313537868875 ], [ 13.385762139383271, 52.507329009449379 ], [ 13.3858148479764, 52.507332082325647 ], [ 13.386650532021926, 52.507380510701005 ], [ 13.386963251269375, 52.507398668477791 ], [ 13.387096201149351, 52.507406396605276 ], [ 13.387920695002668, 52.507454793346973 ], [ 13.387981266559798, 52.507458231600324 ], [ 13.388021902784951, 52.507460542239784 ], [ 13.388136464628188, 52.507467045881114 ], [ 13.388310169922056, 52.507476912726069 ], [ 13.388468327509687, 52.507485894291818 ], [ 13.388488557330414, 52.507487040365355 ], [ 13.388909481851066, 52.507510937511512 ], [ 13.388958981609514, 52.507513750914711 ], [ 13.388967521133111, 52.507514235372597 ], [ 13.389363870770906, 52.507537137937561 ], [ 13.389818672316203, 52.507563391363448 ], [ 13.389953388490817, 52.507571497250638 ], [ 13.389973676913888, 52.507572715079839 ], [ 13.390206478366581, 52.507586723328657 ], [ 13.390301399812438, 52.507592125204383 ], [ 13.390306862194647, 52.507592432673405 ], [ 13.390331170429127, 52.507593820229907 ], [ 13.390387590362518, 52.507597024293894 ], [ 13.39040365355195, 52.507597937054825 ], [ 13.390411780828723, 52.507598402612089 ], [ 13.390494614452493, 52.507603115189106 ], [ 13.390495600774427, 52.507603198036946 ], [ 13.390887284238479, 52.507625915438098 ], [ 13.391731812724926, 52.507674954598535 ], [ 13.39179114722775, 52.507678487296452 ], [ 13.392124483312587, 52.507697840607534 ], [ 13.392451177874673, 52.507719875796596 ], [ 13.392807286070182, 52.507741590792598 ], [ 13.393254707553536, 52.507768876831953 ], [ 13.393404059046858, 52.507778014155555 ], [ 13.393702114910713, 52.507796152164531 ], [ 13.393742691934381, 52.507798631466571 ], [ 13.394058062380957, 52.507817854091869 ], [ 13.394383444673018, 52.507837724591887 ], [ 13.394496223913809, 52.507844685874417 ], [ 13.396058016907903, 52.507940508558058 ], [ 13.396080837912294, 52.507941910007162 ], [ 13.397053058539992, 52.508001548428389 ], [ 13.397363413657262, 52.508022819413569 ], [ 13.398550387754099, 52.5080993428965 ], [ 13.398572398840791, 52.508100751290428 ], [ 13.398771971006491, 52.508113519353365 ], [ 13.398795200226246, 52.508092984899818 ], [ 13.398879283513169, 52.508089929636171 ], [ 13.399150418227244, 52.50808006290827 ], [ 13.399204715886334, 52.508078082434878 ], [ 13.399231525880484, 52.508077109510566 ], [ 13.39923970344589, 52.508087837197024 ], [ 13.399326161077575, 52.508201161986882 ], [ 13.399344579165046, 52.508225290314414 ], [ 13.39935031378057, 52.508232805101066 ], [ 13.399456714453052, 52.508377414337687 ], [ 13.399470447779098, 52.508396096785937 ], [ 13.399601581236315, 52.508574327124713 ], [ 13.399601800581607, 52.508574633088308 ], [ 13.399634781608663, 52.508619449391837 ], [ 13.399694483963673, 52.50870058666009 ], [ 13.399773287925415, 52.508807677735085 ], [ 13.399996333012748, 52.509110951928584 ], [ 13.400189679546504, 52.509375886147289 ], [ 13.400228830474896, 52.509381083417793 ], [ 13.400670760924575, 52.509222507078668 ], [ 13.400909191710721, 52.509140555654298 ], [ 13.401258073055029, 52.509040616477904 ], [ 13.401409856495366, 52.508984326264255 ], [ 13.401554453254223, 52.508930700172598 ], [ 13.402096595706505, 52.508729641011598 ], [ 13.402178196157015, 52.508705325505673 ], [ 13.402179981001657, 52.508704789699522 ], [ 13.402456274263427, 52.508622457664053 ], [ 13.402753429431108, 52.508535680181225 ], [ 13.4030074343167, 52.508433552240696 ], [ 13.403007449181576, 52.508433525309151 ], [ 13.403033366267159, 52.508422790444499 ], [ 13.40320786392421, 52.508350427975152 ], [ 13.403491394147865, 52.508234218917394 ], [ 13.403632339229883, 52.50817652147537 ], [ 13.403820161615338, 52.508092941048815 ], [ 13.40382918070987, 52.50808893206333 ], [ 13.403863397270042, 52.508073701401401 ], [ 13.40391427918507, 52.50805106124659 ], [ 13.404034168682582, 52.507992065562604 ], [ 13.404070858903887, 52.507974017719292 ], [ 13.404299046117305, 52.507861809434708 ], [ 13.404434246099902, 52.507774714223494 ], [ 13.404552277200153, 52.507864918186627 ], [ 13.404580226395426, 52.507886241815783 ], [ 13.404632182513042, 52.507922105798031 ], [ 13.405269546824178, 52.508203887240711 ], [ 13.405284800650678, 52.508210628736215 ], [ 13.405296907925949, 52.508201673829888 ], [ 13.405336541143578, 52.508172352984239 ], [ 13.405372610846895, 52.508145820276127 ], [ 13.405469442689885, 52.508073479917975 ], [ 13.405783270658432, 52.507843651164713 ], [ 13.405903448492499, 52.507755897170817 ], [ 13.406124005966019, 52.507592663480445 ], [ 13.406410190267204, 52.507382012730332 ], [ 13.406601366071207, 52.507232742454683 ], [ 13.407082600370256, 52.506862222549529 ], [ 13.407361283604626, 52.50664769134108 ], [ 13.407394780630824, 52.506664108950545 ], [ 13.407844338808818, 52.50632309295812 ], [ 13.40783956675247, 52.506305380627552 ], [ 13.407877924710457, 52.506280091526108 ], [ 13.407992200461655, 52.506204743891438 ], [ 13.408029435317266, 52.506180189561775 ], [ 13.408054337086813, 52.506189536658972 ], [ 13.408095629325819, 52.506205042860898 ], [ 13.408131510304939, 52.506218858567834 ], [ 13.408162173029366, 52.506230669611945 ], [ 13.408546915546969, 52.5063784983967 ], [ 13.4087587351913, 52.50645989349789 ], [ 13.408986813383089, 52.506547527840397 ], [ 13.409234060838104, 52.506642521011443 ], [ 13.409454996911117, 52.50672740042323 ], [ 13.409951554347597, 52.506918286713947 ], [ 13.409970811094002, 52.506925690957239 ], [ 13.409980505596204, 52.506912956800761 ], [ 13.410155953044111, 52.506682764042623 ], [ 13.410177164913499, 52.50665493557085 ], [ 13.410215067446135, 52.506605228425954 ], [ 13.41021675727386, 52.506603011839942 ], [ 13.410343019416068, 52.506437413980322 ], [ 13.410761614101213, 52.505888291650265 ], [ 13.410983802991986, 52.505596813768854 ], [ 13.411396159508913, 52.505055226458083 ], [ 13.411407216839162, 52.505040697350836 ], [ 13.411430146679148, 52.505010571216587 ], [ 13.411520961508874, 52.504891269186842 ], [ 13.411581930099397, 52.504870342529657 ], [ 13.411582726662655, 52.504870074381323 ], [ 13.411603954277256, 52.504862788932734 ], [ 13.411690148044464, 52.504833209091224 ], [ 13.411711375623526, 52.504825923623279 ], [ 13.411716450186729, 52.504824180465903 ], [ 13.411763566828737, 52.504808009391809 ], [ 13.411810093357278, 52.504792043917924 ], [ 13.412492022993062, 52.504564752415128 ], [ 13.412592298385755, 52.504531324204081 ], [ 13.412623835382325, 52.504520813081747 ], [ 13.412725821661606, 52.504486812713552 ], [ 13.412877177975272, 52.504436402175557 ], [ 13.412910529135111, 52.504425292156128 ], [ 13.412917432412351, 52.50442299508515 ], [ 13.412978101758192, 52.504402768118297 ], [ 13.413018990468716, 52.504389137475158 ], [ 13.413108895341567, 52.504359176879348 ], [ 13.413262124409549, 52.504308077264625 ], [ 13.413290356972974, 52.504298665349573 ], [ 13.4132994728139, 52.504295626363081 ], [ 13.413359655004363, 52.504275569066273 ], [ 13.413534448490354, 52.504217327722948 ], [ 13.413771193557698, 52.504138402741255 ], [ 13.414053221032299, 52.504044470253362 ], [ 13.41407283900692, 52.50403793633901 ], [ 13.4140824301927, 52.504049285617235 ], [ 13.4140975272062, 52.504067142065338 ], [ 13.41417699532392, 52.504161149370297 ], [ 13.414199062558053, 52.504187250006439 ], [ 13.414324341989463, 52.504340129501493 ], [ 13.414410955081211, 52.504442992235759 ], [ 13.414523965218079, 52.504578208896739 ], [ 13.414574365976417, 52.504638230578799 ], [ 13.414610358297974, 52.50468108007226 ], [ 13.414606394747581, 52.504705974702226 ], [ 13.414592287509626, 52.504794457960351 ], [ 13.414632094248173, 52.50480977050352 ], [ 13.414741577691776, 52.504851870968089 ], [ 13.414909213541161, 52.504916344108807 ], [ 13.414939237414607, 52.504908058326521 ], [ 13.415135572186124, 52.504853843202902 ], [ 13.415139627447864, 52.504852727134868 ], [ 13.415313930779959, 52.504804547229618 ], [ 13.415493601422385, 52.504754884880754 ], [ 13.415549313410553, 52.50473948251279 ], [ 13.415796153287999, 52.504671247524335 ], [ 13.415851555373267, 52.504655934327168 ], [ 13.416206101255788, 52.504557938188327 ], [ 13.416210289218066, 52.50455677738691 ], [ 13.416263331611955, 52.504542115817287 ], [ 13.41626579427388, 52.504541428232642 ], [ 13.416319352751293, 52.504526623770957 ], [ 13.416322419977861, 52.504525775502103 ], [ 13.416362161157874, 52.504514792697037 ], [ 13.416414082765087, 52.504500434610165 ], [ 13.416692638883527, 52.504423465266036 ], [ 13.416793650436588, 52.504395516599381 ], [ 13.417074520855449, 52.504317804577624 ], [ 13.417084327051011, 52.504315090038666 ], [ 13.417309515179277, 52.504252780602435 ], [ 13.417319454354971, 52.504249967427214 ], [ 13.417601452244678, 52.50417012633639 ], [ 13.417736978804699, 52.504357097322234 ], [ 13.417808817196848, 52.504456220023428 ], [ 13.417895418724759, 52.504575714284947 ], [ 13.418005167484111, 52.504727115619502 ], [ 13.418096071754778, 52.50485253046741 ], [ 13.418213310635606, 52.50501791889193 ], [ 13.418299074028912, 52.505006836790514 ], [ 13.41834711541626, 52.505000181484149 ], [ 13.418402226097381, 52.504993448738112 ], [ 13.418436551105019, 52.504991082791896 ], [ 13.418518746403336, 52.504987093626951 ], [ 13.418532374918628, 52.504970222889895 ], [ 13.418547219263283, 52.504973430244306 ], [ 13.418620823170778, 52.504989322116977 ], [ 13.418737680279282, 52.505014564094317 ], [ 13.418808178103015, 52.505030207754047 ], [ 13.418843544700103, 52.5050380612278 ], [ 13.418942833190695, 52.505060090898887 ], [ 13.418924477373769, 52.505080098674291 ], [ 13.418990285132189, 52.505115980687158 ], [ 13.419047828995307, 52.505155442780492 ], [ 13.419102259562267, 52.505199302792768 ], [ 13.419120299132086, 52.505219122669153 ], [ 13.419135204169281, 52.505239889931723 ], [ 13.419207685347384, 52.505344216107474 ], [ 13.419298080506634, 52.505466151330353 ], [ 13.419424035342125, 52.505639641708974 ], [ 13.419430912540886, 52.505649116592885 ], [ 13.419445436196735, 52.505645421602253 ], [ 13.419534612906894, 52.50562272497725 ], [ 13.419669808043702, 52.50558518215486 ], [ 13.419851232807231, 52.505534806333593 ], [ 13.420008827117167, 52.505491038654618 ], [ 13.420322033601163, 52.505402016428704 ], [ 13.420329214984315, 52.505400033940333 ], [ 13.42036282157842, 52.505390737552489 ], [ 13.420537180275401, 52.505342541074555 ], [ 13.420902860691477, 52.505243292170242 ], [ 13.421132300487521, 52.505175474114878 ], [ 13.421233930961087, 52.505150362489765 ], [ 13.42138946469538, 52.505115225467343 ], [ 13.421542217232881, 52.505076398932601 ], [ 13.421810979925116, 52.505042607909203 ], [ 13.421860299863084, 52.505039592943533 ], [ 13.421891117473125, 52.505037705196877 ], [ 13.421977206365026, 52.505032444279088 ], [ 13.422045396932811, 52.505028266674259 ], [ 13.422095011459216, 52.505025234147176 ], [ 13.422167238949054, 52.505018464989902 ], [ 13.422239639494526, 52.505012423357044 ], [ 13.42231211918287, 52.505006746094757 ], [ 13.422384742394737, 52.505001798745724 ], [ 13.422457429961923, 52.504997216850057 ], [ 13.422530231432813, 52.504993366810091 ], [ 13.422603082432151, 52.504989883082843 ], [ 13.422676017645765, 52.504987132706226 ], [ 13.422748987530692, 52.504984749277696 ], [ 13.422822011889002, 52.504983100246299 ], [ 13.422895056040312, 52.504981818573995 ], [ 13.422968124890817, 52.504981271896256 ], [ 13.423041198643915, 52.504981092763956 ], [ 13.423114267306916, 52.504981648774226 ], [ 13.423187325979111, 52.504982572291567 ], [ 13.423260349775029, 52.504984230649946 ], [ 13.423333348692834, 52.504986256252309 ], [ 13.42340628296949, 52.504989015944837 ], [ 13.423479177495924, 52.504992142393718 ], [ 13.423551977655835, 52.504996001732984 ], [ 13.423624723217694, 52.50500022711671 ], [ 13.423697344745317, 52.505005183742909 ], [ 13.423769896860412, 52.505010505477856 ], [ 13.423842295349356, 52.505016556360204 ], [ 13.423914609653687, 52.505022971192361 ], [ 13.423986740833829, 52.505030112630955 ], [ 13.424058773108699, 52.505037616637878 ], [ 13.424130592873244, 52.505045844265993 ], [ 13.424202299072283, 52.505054432859232 ], [ 13.424273763504718, 52.505063741645934 ], [ 13.42435404854386, 52.50507477067049 ], [ 13.424407956991873, 52.505082521102054 ], [ 13.424461702535428, 52.50509068196903 ], [ 13.424515363488386, 52.505099047568592 ], [ 13.424568848959954, 52.505107821668453 ], [ 13.424622243568214, 52.505116799497515 ], [ 13.424675450179755, 52.505126183748217 ], [ 13.424744677767546, 52.505138722551365 ], [ 13.424813561422457, 52.505151947904814 ], [ 13.424882266268957, 52.505165515210848 ], [ 13.424950599793574, 52.505179763733842 ], [ 13.425137003269029, 52.505227269794581 ], [ 13.425315846208141, 52.505272857544639 ], [ 13.425497208558498, 52.505321513721263 ], [ 13.425820881311989, 52.505424600201501 ], [ 13.426422318623537, 52.505677917111306 ], [ 13.426446087002715, 52.505687923422784 ], [ 13.426501120272883, 52.505711102994084 ], [ 13.42668281571679, 52.505787634183832 ], [ 13.426708847865671, 52.505798596843157 ], [ 13.426730419330944, 52.505792971776629 ], [ 13.426979707196487, 52.50572797075224 ], [ 13.42720030648004, 52.505666066126551 ], [ 13.427209407756253, 52.505679102980167 ], [ 13.427246383650576, 52.505732051152371 ], [ 13.427356287430685, 52.505889465010242 ], [ 13.427440132181099, 52.506009396921648 ], [ 13.427537731604046, 52.506149122328466 ], [ 13.427617918263362, 52.506264015483985 ], [ 13.427628454085012, 52.506279040744545 ], [ 13.427675235101566, 52.506345988340371 ], [ 13.427690189867086, 52.506367383503409 ], [ 13.427739327032649, 52.506437687017694 ], [ 13.427784864635049, 52.506502844154461 ], [ 13.427799526866567, 52.506523816443561 ], [ 13.42784633773498, 52.506590790978407 ], [ 13.427954475367327, 52.50674563113428 ], [ 13.428062614163402, 52.506900381323234 ], [ 13.428251031509081, 52.507170141094576 ], [ 13.428266454917598, 52.507192246904516 ], [ 13.428429669532733, 52.507424067845506 ], [ 13.428772615518573, 52.50791685305569 ], [ 13.428784527176967, 52.507933965388325 ], [ 13.428816442823047, 52.507979823185785 ], [ 13.429027819270784, 52.50803543476205 ], [ 13.429028360814536, 52.508036190536473 ], [ 13.429037228965329, 52.508048894351973 ], [ 13.429393128289197, 52.508559088303393 ], [ 13.42940211370966, 52.508571963030143 ], [ 13.429402289278976, 52.508572223932646 ], [ 13.429240115534826, 52.508671254714969 ], [ 13.429206100393488, 52.508725548560392 ], [ 13.429200378296471, 52.5087535141191 ], [ 13.429193978519438, 52.508784803567224 ], [ 13.429186900938603, 52.508819443864077 ], [ 13.429184989320525, 52.508823852612608 ], [ 13.429182427129989, 52.50882813670988 ], [ 13.42917954725273, 52.508832345016131 ], [ 13.429176050772142, 52.508836375661467 ], [ 13.42917225653016, 52.508840304769784 ], [ 13.429167891193375, 52.5088440065323 ], [ 13.429163253592979, 52.508847582911997 ], [ 13.429158101094036, 52.508850886515575 ], [ 13.429152706924558, 52.508854043236994 ], [ 13.429146863688858, 52.50885688685684 ], [ 13.429122519013639, 52.508866488769783 ], [ 13.429136398509362, 52.5088849703044 ], [ 13.429136706655569, 52.508886088970442 ], [ 13.429136855686433, 52.508887219676282 ], [ 13.429136924654644, 52.508888353251848 ], [ 13.42913683329154, 52.508889486237649 ], [ 13.429136661928297, 52.508890615764948 ], [ 13.429136331699135, 52.508891732082745 ], [ 13.429135922870259, 52.508892838671656 ], [ 13.429135359297726, 52.508893919652273 ], [ 13.429134719840395, 52.508894984796356 ], [ 13.42913393234959, 52.508896012361568 ], [ 13.429133072957997, 52.508897018247417 ], [ 13.429132074718806, 52.508897975212328 ], [ 13.429131009765305, 52.508898905017624 ], [ 13.429129817472163, 52.508899775378055 ], [ 13.429128564766916, 52.508900613552811 ], [ 13.429127198359598, 52.508901382752924 ], [ 13.429125778852709, 52.508902115279589 ], [ 13.42912426118288, 52.50890277045535 ], [ 13.429122698614043, 52.508903385083201 ], [ 13.429121055063064, 52.508903915277514 ], [ 13.429020585215033, 52.508935317196595 ], [ 13.428920719926158, 52.508966531323381 ], [ 13.428918914343765, 52.508967257428672 ], [ 13.428917004168992, 52.508967875928235 ], [ 13.428915048075263, 52.50896843846445 ], [ 13.428913013455531, 52.508968885432395 ], [ 13.428910946387713, 52.508969273029365 ], [ 13.42890882844746, 52.508969539431419 ], [ 13.428906692185842, 52.508969744254891 ], [ 13.428904533733618, 52.508969824706455 ], [ 13.428902371457072, 52.508969842616459 ], [ 13.428900216120271, 52.508969735491412 ], [ 13.42889807153349, 52.508969566125963 ], [ 13.428895962876537, 52.508969273589571 ], [ 13.428893879326665, 52.50896892037192 ], [ 13.42889185997047, 52.508968448337008 ], [ 13.428889879570935, 52.508967918406533 ], [ 13.428887990331427, 52.508967276414033 ], [ 13.428887356011076, 52.508967500017192 ], [ 13.428856231130593, 52.50897823183108 ], [ 13.428666176631163, 52.508991483580985 ], [ 13.428377282968299, 52.509072867131458 ], [ 13.428372726124289, 52.509070882454274 ], [ 13.428163145898615, 52.5092269058201 ], [ 13.428081881935876, 52.509287374684618 ], [ 13.426707315191262, 52.510094117790381 ], [ 13.426020729232567, 52.510497053155987 ], [ 13.426015138443214, 52.510492469524607 ], [ 13.425390602877645, 52.510857771953454 ], [ 13.425110164698618, 52.511021811112577 ], [ 13.424873585448639, 52.511163790571686 ], [ 13.424520979943129, 52.511367578492404 ], [ 13.424403832347631, 52.511430068375212 ], [ 13.423655011019569, 52.511808853531029 ], [ 13.42341719181133, 52.511946289735647 ], [ 13.423248104884083, 52.512044006328843 ], [ 13.423184432267988, 52.512088155539729 ], [ 13.423078038712205, 52.512130353022314 ], [ 13.423051050755538, 52.512127988050892 ], [ 13.423032055228457, 52.5121172793488 ], [ 13.423011419140689, 52.512126895371082 ], [ 13.422780583775907, 52.512234435111537 ], [ 13.422784144348558, 52.512238467216967 ], [ 13.422793917556721, 52.512249546488135 ], [ 13.422919320895263, 52.512380651456283 ], [ 13.422939505672071, 52.512401759649066 ], [ 13.423021227184545, 52.512487200568188 ], [ 13.423093512835639, 52.512565094474517 ], [ 13.423110572924752, 52.512583492312643 ], [ 13.423219456257568, 52.512700837099324 ], [ 13.423232031664343, 52.512714374476879 ], [ 13.423241441130344, 52.512724527471157 ], [ 13.423245486361461, 52.512728883911009 ], [ 13.423239467331374, 52.512730913505933 ], [ 13.423248510434714, 52.51274066147549 ], [ 13.423265277727822, 52.512758699325737 ], [ 13.423277853123642, 52.512772245684758 ], [ 13.423298269896943, 52.512794225907697 ], [ 13.423310845313026, 52.51280777226318 ], [ 13.423323933778269, 52.512821858691353 ], [ 13.42333652389808, 52.512835414055843 ], [ 13.423348703633462, 52.512848528373333 ], [ 13.423358523630259, 52.512859104430795 ], [ 13.423386415434212, 52.512889131448112 ], [ 13.423417411424843, 52.512919981566455 ], [ 13.423421986040621, 52.512924536613383 ], [ 13.423453427209022, 52.51295787675781 ], [ 13.423696281049773, 52.513215317601173 ], [ 13.423735608688483, 52.513257019691302 ], [ 13.42382417324003, 52.513350901150346 ], [ 13.423827266126693, 52.513354177552571 ], [ 13.423832601679978, 52.51335983923186 ], [ 13.423840912865652, 52.513368651300226 ], [ 13.423947099398289, 52.513563904049995 ], [ 13.423968146919496, 52.513602609179515 ], [ 13.424010943228025, 52.513681296715866 ], [ 13.424064112608425, 52.513740105807699 ], [ 13.42407408830516, 52.513752137889476 ], [ 13.424102623869883, 52.513786542278261 ], [ 13.424172160591525, 52.513867306490731 ], [ 13.42423904567759, 52.513944974755759 ], [ 13.424298969019077, 52.514021489784092 ], [ 13.424389689216319, 52.514137310626893 ], [ 13.424410835369462, 52.51416430638595 ], [ 13.424859158295185, 52.514752530750172 ], [ 13.424888119110205, 52.51479084481084 ], [ 13.425004035947925, 52.514944226916562 ], [ 13.42509614351501, 52.515069925763008 ], [ 13.42510317539587, 52.515081593278083 ], [ 13.425139782250614, 52.51514231474389 ], [ 13.425396574286712, 52.515587531374656 ], [ 13.425483736704431, 52.515738617434366 ], [ 13.425533953225615, 52.515825674481121 ], [ 13.425562759068061, 52.515875625665593 ], [ 13.425598514566726, 52.51593919426427 ], [ 13.425670389214179, 52.516067410428597 ], [ 13.425671981635343, 52.51607030679741 ], [ 13.425675341686164, 52.516076441320777 ], [ 13.425731002839106, 52.516177832060499 ], [ 13.425733910071846, 52.516183130065578 ], [ 13.425819243071425, 52.516338571222484 ], [ 13.425819564440092, 52.516339164879184 ], [ 13.425823348236326, 52.516346063972662 ], [ 13.425831135070515, 52.516360239965735 ], [ 13.425871544738101, 52.516433836359724 ], [ 13.425959085104312, 52.516593217255675 ], [ 13.426000181318559, 52.516668180719975 ], [ 13.426046707135706, 52.516747745493511 ], [ 13.426188244061702, 52.516989839926211 ], [ 13.426250703281218, 52.517096660801151 ], [ 13.426370488812017, 52.517311929304924 ], [ 13.42638399001906, 52.517336179847199 ], [ 13.426388027726437, 52.51734235144216 ], [ 13.42640365193043, 52.517366228141448 ], [ 13.426410205920462, 52.517376241243291 ], [ 13.426456435067294, 52.517446863745157 ], [ 13.426475760633638, 52.517476390238897 ], [ 13.426528363377184, 52.517567602649777 ], [ 13.426702571958504, 52.517869736778579 ], [ 13.426732024180554, 52.517920668283409 ], [ 13.426754137804394, 52.517959167996999 ], [ 13.426511270055657, 52.51800945938583 ], [ 13.426505829096623, 52.518010582483946 ], [ 13.426505784821327, 52.518010600382112 ], [ 13.426338574242621, 52.518045221066302 ], [ 13.426331531902337, 52.518032439283679 ], [ 13.426151203200654, 52.518068601146673 ], [ 13.426189796794977, 52.51812414938022 ], [ 13.426207051957926, 52.518148990424343 ], [ 13.426240362232958, 52.518196945058229 ], [ 13.426305379938118, 52.518289032307301 ], [ 13.42592691860073, 52.518391735886311 ], [ 13.426190547252965, 52.518751994502466 ], [ 13.426446684272083, 52.519101986200582 ], [ 13.42645719623231, 52.519116364460508 ], [ 13.42646126639506, 52.519121916035495 ], [ 13.426477034433946, 52.519143465450114 ], [ 13.426487546454807, 52.519157834720772 ], [ 13.426490899190277, 52.51916241453587 ], [ 13.426529682546978, 52.519215410794452 ], [ 13.426568861054969, 52.519269000818888 ], [ 13.426729207991178, 52.519488102820389 ], [ 13.426881152561535, 52.519695714612851 ], [ 13.426916525471091, 52.519744049959485 ], [ 13.427354708465014, 52.519627415883093 ], [ 13.427372919814125, 52.519622696297368 ], [ 13.427391596901515, 52.519618711139302 ], [ 13.427410477176371, 52.519615099792794 ], [ 13.427429703301584, 52.519612247122943 ], [ 13.427463377873142, 52.519608211211548 ], [ 13.427497584090043, 52.519606520166597 ], [ 13.428401342936455, 52.519585904237182 ], [ 13.428423939055779, 52.519632267645854 ], [ 13.428456840686906, 52.519699784625473 ], [ 13.428906496480721, 52.520622536648077 ], [ 13.428915401063303, 52.520640803163921 ], [ 13.428932801384585, 52.52067652672298 ], [ 13.428940903149876, 52.520693138374455 ], [ 13.429189767078501, 52.521203826847227 ], [ 13.429181455732236, 52.521224041641517 ], [ 13.428896308241823, 52.521345871957251 ], [ 13.428844970321267, 52.521367803119261 ], [ 13.428608649821046, 52.521468775764305 ], [ 13.428567393267651, 52.521486401162598 ], [ 13.428565459567658, 52.521487233671785 ], [ 13.428050096683375, 52.521707411212297 ], [ 13.427787955433855, 52.521819410265771 ], [ 13.427622477656644, 52.521894663736937 ], [ 13.427050246626456, 52.522138735251779 ], [ 13.426719357617973, 52.522279867063261 ], [ 13.425931105490488, 52.522623758654561 ], [ 13.425547876778465, 52.522790949157063 ], [ 13.425750482780861, 52.522813759673078 ], [ 13.425908355260967, 52.522831111023194 ], [ 13.426004042987424, 52.522925416147586 ], [ 13.426125233859047, 52.522929180277345 ], [ 13.426154225022772, 52.522929903399735 ], [ 13.426521328567871, 52.522939080108699 ], [ 13.426514407633421, 52.523064673020016 ], [ 13.426508723080445, 52.523167837661489 ], [ 13.426501915173333, 52.523291238047207 ], [ 13.426418555018168, 52.523293638387472 ], [ 13.426335615810892, 52.52329927812098 ], [ 13.426254502264213, 52.523306410597947 ], [ 13.426173649241603, 52.523314569763038 ], [ 13.426092939571397, 52.523323240956103 ], [ 13.426012542762701, 52.523332933385134 ], [ 13.425881931873276, 52.523352760748509 ], [ 13.425770237287704, 52.523372781844259 ], [ 13.425650801830937, 52.523403771213275 ], [ 13.425484798482369, 52.523446849042074 ], [ 13.425433849527082, 52.523440795352748 ], [ 13.425052801215617, 52.523551354885825 ], [ 13.424778841806955, 52.523630840839289 ], [ 13.424333430083673, 52.523768211834714 ], [ 13.424246081488887, 52.523796010515959 ], [ 13.424173779974272, 52.523819620206261 ], [ 13.424098968412617, 52.523844384816378 ], [ 13.424027834264267, 52.523870413790114 ], [ 13.42395596006299, 52.523897160374858 ], [ 13.423859308160518, 52.52393268022616 ], [ 13.42379508182133, 52.523955908118644 ], [ 13.423769285013758, 52.523965236827024 ], [ 13.423703014501545, 52.523990824619162 ], [ 13.423656666064186, 52.52400943707039 ], [ 13.423376373304446, 52.524119229429616 ], [ 13.421632270245128, 52.524805171637141 ], [ 13.421476737345388, 52.52486633387462 ], [ 13.421476678369544, 52.52486634275887 ], [ 13.421350681109208, 52.52491590951572 ], [ 13.4205402094939, 52.52523464517639 ], [ 13.420514569865535, 52.525244710342719 ], [ 13.420501196378929, 52.525249989088366 ], [ 13.420501137444466, 52.525249988985685 ], [ 13.420471113620668, 52.52526181684793 ], [ 13.419795878967498, 52.525528742058867 ], [ 13.419753218923884, 52.525545607056067 ], [ 13.419488766516768, 52.525657197222628 ], [ 13.418376259653952, 52.526126586995041 ], [ 13.418249514779198, 52.526180067250827 ], [ 13.417963396928432, 52.526306919887872 ], [ 13.417849321125686, 52.526357555362985 ], [ 13.417691820463476, 52.526427471004808 ], [ 13.417449832246707, 52.526534881607645 ], [ 13.417288934255426, 52.526606309426754 ], [ 13.417280636139891, 52.52660998820209 ], [ 13.417215978516541, 52.526638693533364 ], [ 13.416889987178783, 52.52678339210015 ], [ 13.416904690304234, 52.526796107128405 ], [ 13.416707443515239, 52.526885541600642 ], [ 13.416685167047344, 52.526875904470991 ], [ 13.416490486705714, 52.52696152386536 ]];
-
-                    });
-                    adhHttp.options(value).then((options : AdhHttp.IOptions) => {
-                        scope.options = options;
-                    });
-                }
-            });
-        }
-    };
-};
 
 
 export var moduleName = "adhMeinBerlinKiezkassen";
 
 export var register = (angular) => {
+    AdhMeinBerlinKiezkassenProcess.register(angular);
     AdhMeinBerlinKiezkassenProposal.register(angular);
 
     angular
         .module(moduleName, [
-            AdhTabs.moduleName,
+            AdhMeinBerlinKiezkassenProcess.moduleName,
             AdhMeinBerlinKiezkassenProposal.moduleName
-        ])
-        .directive("adhMeinBerlinKiezkassenDetail", ["adhConfig", "adhHttp", detailDirective]);
+        ]);
 };
